Fixed warnings

+++ GameEngine/GameLogic/Character/ProjectileSystem/Projectile/Projectile.cpp
@@ -64,8 +64,8 @@ void Projectile::Update(double deltatime)
 	if(isFired == true)
 	{
 
-		age += deltatime;
-		position = position+(direction*speed*deltatime);
+		age += (float)deltatime;
+		position = position+(direction*speed*(float)deltatime);
 		renderInfo = { position, direction };
 
 	}
